Store breakpoints inside collab db when you change them

+++ crates/collab/migrations.sqlite/20221109000000_test_schema.sql
@@ -442,7 +442,7 @@ CREATE TABLE IF NOT EXISTS "breakpoints" (
     "project_id" INTEGER NOT NULL REFERENCES projects (id) ON DELETE CASCADE,
     "position" INTEGER NOT NULL,
     "log_message" TEXT NULL,
-    "worktree_id" INTEGER NOT NULL,
+    "worktree_id" BIGINT NOT NULL,
     "path" TEXT NOT NULL,
     "kind" VARCHAR NOT NULL
 );
